Refactor and updated gitignore
