+++ .gitignore
@@ -1,2 +1,5 @@
 venv/
 data/
+models/
+
+__pycache__/
