+++ src/train.py
@@ -16,6 +16,9 @@ from tqdm import tqdm
 
 from PIL import Image
 
+from RetinaDiseaseDataset import RetinaDiseaseDataset
+from RetinaDiseaseClassifier import RetinaDiseaseClassifier
+
 csv_path='../data/ocular-disease-recognition-odir5k/full_df.csv'
 csv = pd.read_csv(csv_path)
 
@@ -27,45 +30,7 @@ data_transform = transforms.Compose([transforms.Resize((IMAGE_SIZE, IMAGE_SIZE))
 
 diagnosis_list = ['Normal', 'Diabetes', 'Glaucoma', 'Cataract', 'Age', 'Hypertension', 'Pathological Myopia', 'Other']
 
-class RetinaDiseaseDataset(Dataset):
-    def __init__(self, csv_file, root_dir, transform=None):
-        self.df = pd.read_csv(csv_file) 
-        self.transform = transform 
-        self.root_dir = root_dir
-        
-        # Define a mapping from string labels to integers
-        self.label_mapping = {
-            "N": 0,  # Normal
-            "D": 1,  # Diabetes
-            "G": 2,  # Glaucoma
-            "C": 3,  # Cataract
-            "A": 4,  # Age-related Macular Degeneration
-            "H": 5,  # Hypertension
-            "M": 6,  # Pathological Myopia
-            "O": 7   # Other diseases/abnormalities
-        }
 
-    def __len__(self):
-        return len(self.df)  
-
-    def __getitem__(self, index):
-        image_path = os.path.join(self.root_dir,
-                                self.df.filename[index])
-        image = Image.open(image_path)
-        
-        label = self.df.labels[index]
-        label = label.strip("[]").strip("'\"")  # Remove brackets from cell value
-        
-        # Convert the label to int
-        label = self.label_mapping.get(label, -1)  # Use -1 for any unknown labels
-        
-        # Convert label to a tensor
-        label = torch.tensor(label, dtype=torch.long)
-        
-        if self.transform:
-            image = self.transform(image)
-        
-        return image, label
 
 dataset = RetinaDiseaseDataset(
     csv_path,
@@ -73,24 +38,7 @@ dataset = RetinaDiseaseDataset(
     data_transform
 )
     
-class RetinaDiseaseClassifier(nn.Module):
-    def __init__(self, num_classes=8):
-        super(RetinaDiseaseClassifier, self).__init__()
-        # Where we define all the parts of the model
-        self.base_model = timm.create_model('resnet50', pretrained=True)
-        self.features = nn.Sequential(*list(self.base_model.children())[:-1])
 
-        out_size = 2048
-        self.classifier = nn.Sequential(
-            nn.Flatten(),
-            nn.Linear(out_size, num_classes)
-        )
-    
-    def forward(self, x):
-        x = self.features(x)
-        output = self.classifier(x)
-        return output
-    
 
 # Calculate the lengths for the splits
 total_size = len(dataset)
@@ -112,7 +60,7 @@ model = RetinaDiseaseClassifier(num_classes=8)
 model.load_state_dict(torch.load('../models/retinai_resnet50_0.0.1.pth'))
 
 # Simple training loop
-num_epochs = 3
+num_epochs = 1
 train_losses, val_losses = [], []
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
